# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Portfolio Service API",
    "description": "Declarative Zod Validated & Programmatically Generated API Spec",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/blog": {
      "get": {
        "summary": "List unpublished blog drafts",
        "description": "Returns persisted unpublished BlogPost rows for authorized administrators only. Public fallback content is never used for this inventory.",
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10000,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated draft collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogDraftCollection"
                }
              }
            }
          },
          "400": {
            "description": "Invalid bounded pagination request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Administrator access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unable to load drafts from persistence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an unpublished sanitized blog draft",
        "description": "Creates a persisted BlogPost in server-owned unpublished state for an authorized administrator. Publication, identity, and timestamp fields are rejected.",
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogDraftCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sanitized unpublished draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogDraftCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body, taxonomy, field bounds, or mass-assignment attempt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Administrator access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "A draft with the slug already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "500": {
            "description": "Unable to create draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/blog/{id}": {
      "get": {
        "summary": "Read an unpublished blog draft",
        "description": "Returns one persisted unpublished BlogPost row for an authorized administrator. Published records and public fallback content are not available through this endpoint.",
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 191
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated unpublished draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogDraftReadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid draft identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Administrator access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unpublished draft not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unable to load draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Edit an unpublished sanitized blog draft",
        "description": "Partially updates a persisted BlogPost only while it remains unpublished. Publication, identity, and timestamp fields are rejected, changed HTML is sanitized, and old and new slug caches are invalidated only after persistence succeeds.",
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 191
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogDraftUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sanitized unpublished draft updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogDraftUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body, empty edit, taxonomy, field bounds, or mass-assignment attempt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Administrator access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unpublished draft not found or no longer editable after concurrent publication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "A draft with the requested slug already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "500": {
            "description": "Unable to update draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a blog post or draft",
        "description": "Deletes a persisted BlogPost or draft record by ID and evicts associated caches.",
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 191
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog post or draft deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid draft identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Administrator access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Blog draft not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unable to delete blog post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ClerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk session token or administrative JWT"
      },
      "CronSecretAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Secret",
        "description": "Cron secret bearer authorization header"
      }
    },
    "schemas": {
...
      "BlogDraftCreate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 180
          },
          "slug": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "dek": {
            "type": "string",
            "minLength": 10,
            "maxLength": 500
          },
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50000
          },
          "pillar": {
            "type": "string",
            "enum": [
              "clinical-data-engineering",
              "formal-verification",
              "accessibility-engineering",
              "browser-graphics-engineering",
              "agent-first-dx",
              "field-notes"
            ]
          },
          "tags": {
            "minItems": 1,
            "maxItems": 12,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 50
            }
          },
          "heroImageUrl": {
            "type": "string",
            "maxLength": 2048,
            "format": "uri",
            "nullable": true
          }
        },
        "required": [
          "title",
          "slug",
          "dek",
          "body",
          "pillar",
          "tags"
        ],
        "additionalProperties": false
      },
      "BlogDraftUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 180
          },
          "slug": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "dek": {
            "type": "string",
            "minLength": 10,
            "maxLength": 500
          },
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50000
          },
          "pillar": {
            "type": "string",
            "enum": [
              "clinical-data-engineering",
              "formal-verification",
              "accessibility-engineering",
              "browser-graphics-engineering",
              "agent-first-dx",
              "field-notes"
            ]
          },
          "tags": {
            "minItems": 1,
            "maxItems": 12,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 50
            }
          },
          "heroImageUrl": {
            "type": "string",
            "maxLength": 2048,
            "format": "uri",
            "nullable": true
          },
          "published": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "BlogDraftCollection": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "slug": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "dek": {
                  "type": "string"
                },
                "body": {
                  "type": "string"
                },
                "pillar": {
                  "type": "string"
                },
                "tags": {
                  "type": "string"
                },
                "published": {
                  "type": "boolean",
                  "const": false
                },
                "reading_time_minutes": {
                  "type": "integer",
                  "minimum": 1,
                  "nullable": true
                },
                "hero_image_url": {
                  "type": "string",
                  "format": "uri",
                  "nullable": true
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "id",
                "slug",
                "title",
                "dek",
                "body",
                "pillar",
                "tags",
                "published",
                "created_at",
                "updated_at"
              ],
              "additionalProperties": false
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "minimum": 1
              },
              "pageSize": {
                "type": "integer",
                "minimum": 1,
                "maximum": 100
              },
              "total": {
                "type": "integer",
                "minimum": 0
              }
            },
            "required": [
              "page",
              "pageSize",
              "total"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "additionalProperties": false
      },
      "BlogDraftCreateResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "dek": {
                "type": "string"
              },
              "body": {
                "type": "string"
              },
              "pillar": {
                "type": "string"
              },
              "tags": {
                "type": "string"
              },
              "published": {
                "type": "boolean",
                "const": false
              },
              "reading_time_minutes": {
                "type": "integer",
                "minimum": 1,
                "nullable": true
              },
              "hero_image_url": {
                "type": "string",
                "format": "uri",
                "nullable": true
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "id",
              "slug",
              "title",
              "dek",
              "body",
              "pillar",
              "tags",
              "published",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "success",
          "data"
        ],
        "additionalProperties": false
      },
      "BlogDraftReadResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "dek": {
                "type": "string"
              },
              "body": {
                "type": "string"
              },
              "pillar": {
                "type": "string"
              },
              "tags": {
                "type": "string"
              },
              "published": {
                "type": "boolean",
                "const": false
              },
              "reading_time_minutes": {
                "type": "integer",
                "minimum": 1,
                "nullable": true
              },
              "hero_image_url": {
                "type": "string",
                "format": "uri",
                "nullable": true
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "id",
              "slug",
              "title",
              "dek",
              "body",
              "pillar",
              "tags",
              "published",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
      "BlogDraftUpdateResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "dek": {
                "type": "string"
              },
              "body": {
                "type": "string"
              },
              "pillar": {
                "type": "string"
              },
              "tags": {
                "type": "string"
              },
              "published": {
                "type": "boolean",
                "const": false
              },
              "reading_time_minutes": {
                "type": "integer",
                "minimum": 1,
                "nullable": true
              },
              "hero_image_url": {
                "type": "string",
                "format": "uri",
                "nullable": true
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "id",
              "slug",
              "title",
              "dek",
              "body",
              "pillar",
              "tags",
              "published",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "success",
          "data"
        ],
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}